# Conflicts:
#	backend/src/main/java/com/akademikplus/akademik_plus/entity/Room.java

+++ backend/src/main/java/com/akademikplus/akademik_plus/repository/MaintenanceRequestRepository.java
@@ -3,5 +3,5 @@ package com.akademikplus.akademik_plus.repository;
 import com.akademikplus.akademik_plus.entity.MaintenanceRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MaintenanceRequestRepository extends JpaRepository<MaintenanceRequest, Integer> {
+public interface MaintenanceRequestRepository extends JpaRepository<MaintenanceRequest, Long> {
 }
